+++ biopipen/reports/scrna/TopExpressingGenes.svelte
@@ -1,0 +1,43 @@
+{% from "utils/misc.liq" import report_jobs -%}
+<script>
+    import { Image, DataTable } from "$libs";
+    import { Tabs, Tab, TabContent, InlineNotification } from "$ccs";
+</script>
+
+{%- macro report_job(job, h=1) -%}
+  {% for casedir in job.out.outdir | glob: "*" %}
+    {% set ident = casedir | basename %}
+    <h{{h}}>{{ident}}</h{{h}}>
+    <h{{h+1}}>Top expression genes</h{{h+1}}>
+    <DataTable
+        src={{ casedir | joinpaths: "expr.txt" | quote }}
+        data={ {{ casedir | joinpaths: "expr.txt" | datatable: sep="\t", nrows=envs.n }} }
+        />
+
+    <h{{h+1}}>Enrichment analysis</h{{h+1}}>
+    <Tabs>
+        {% for enrtxt in casedir | glob: "Enrichr-*.txt"  %}
+        {%  set db = enrtxt | stem | replace: "Enrichr-", "" %}
+        <Tab label="{{db}}" title="{{db}}" />
+        {% endfor %}
+        <div slot="content">
+            {% for enrtxt in casedir | glob: "Enrichr-*.txt" %}
+            {%  set db = enrtxt | stem | replace: "Enrichr-", "" %}
+            <TabContent>
+                <Image src={{casedir | joinpaths: "Enrichr-" + db + ".png" | quote}} />
+                <DataTable
+                    src={{ enrtxt | quote }}
+                    data={ {{ enrtxt | datatable: sep="\t", nrows=100 }} }
+                    />
+            </TabContent>
+            {% endfor %}
+        </div>
+    </Tabs>
+  {% endfor %}
+{%- endmacro -%}
+
+{%- macro head_job(job) -%}
+  <h1>{{job.in.srtobj | stem | escape}}</h1>
+{%- endmacro -%}
+
+{{ report_jobs(jobs, head_job, report_job) }}